requirements.txt
streamlit>=1.37
requests>=2.32
python-dotenv>=1.0
pytest>=8.0
openai>=1.0.0
anthropic>=0.30.0
google-generativeai>=0.8.0
